+++ gframe/menu_handler.cpp
@@ -497,7 +497,7 @@ bool MenuHandler::OnEvent(const irr::SEvent& event) {
 				if(players.size() > decks.size())
 					break;
 				auto replay_name = Utils::GetFileName(ReplayMode::cur_replay.GetReplayName());
-				for(int i = 0; i < decks.size(); i++) {
+				for(size_t i = 0; i < decks.size(); i++) {
 					gdeckManager->SaveDeck(sanitize(fmt::format(EPRO_TEXT("{} player{:02} {}"), replay_name, i, Utils::ToPathString(players[i]))), decks[i].main_deck, decks[i].extra_deck, std::vector<int>());
 				}
 				mainGame->stACMessage->setText(gDataManager->GetSysString(1367).c_str());
